Refactor scatter-in-cluster rule
To fix BTS-35, i.e. to make scatter-in-cluster work with smart joins that
are dependent on nodes which are run on a coordinator, we now insert a normal
SCATTER/GATHER snippet around the southmost node in a smartjoined block of,
and subsequently move the SCATTER part upwards past all nodes involved in the
smart join.

While we're here we delete the scatter-iresearch-view-in-cluster rule, because
it was essentially a duplicate of scatter-in-cluster, just for iresearch views.