don't pessimize view usage in AQL queries
previously, when an arangosearch view was used in a cluster AQL query,
the lookup was calling the lookup for collections first, which would not
find a collection of the request name, which would trigger a plan reload
and then try again, only to not find the collection once more.
only after that the actual view lookup happened, which would happily
return the view object.
that old algorithm put an unnecessary delayed on the usage of views in
cluster AQL queries, which should be eliminated by this PR.