Fix bug where early abort condition aborts too early
Before this path the k-shortest-paths algorithm kept a record
of the weight of the done vertex furthest away from the centre,
and a record of the weight of the currently best path found.

This was used to stop search if

  start.furthest + end.furthest > currentBest

which is wrong, as (wlog) the start side can find *any* vertex in the
target side, including the target vertex itself.

We now stop search *for a neighbourhood*, if the known highest weight
is higher than current best path weight, as we are not
going to find a shorter path originating in this neighbourhood.