Fix an arangod configuration bug
where the server tried to interpret
a directory as conf-file leading to
an early exit. Now the path is instead
ignored if it does not point to a
regular file.

This was especially a problem when
the original path was empty and
arangod tried to find a local
override. In this case the search
had '.local' as candidate witch is
a common directory in user homes.