fix unprotected access to _computedValue (#11655)
AstNodes have a member `_computedValue` which can contain a pointer to
dynamically allocated memory. this memory is owned by the AstNode and is
freed when the AstNode is destroyed. this works fine for AstNodes that
are dynamically created and destroyed in a query.
However, there were some singleton AstNodes which were created just once
for the entire server process and were shared between all queries:
  * NullNode
  * FalseNode
  * TrueNode
  * ZeroNode
  * EmptyStringNode
The initialization of the `_computedValue` attribute for these singleton
nodes was not thread-safe, so it multiple queries computed the value for
the same singleton node simultaneously, this could have caused races and
a (small) one-time memory leak.

The fix now makes nodes per-query singleton nodes, so they can be reused
within the same query, but are not shared between queries. in addition,
the singleton nodes are all initialized in a query before there is any
parallelization. as a bonus, no dynamic memory allocation is now needed
for the `_computedValues` of these per-query singleton nodes.