make maxMemoryUsage non-atomic (#11977)
maxMemoryUsage seems to be set only once, but in a situation when
there is yet no concurrency for the query.
The ResourceUsage object is owned by the QueryContext and created
in the QueryContext's constructor. Query is derived from QueryContext
and calls `setMemoryLimit` in ResourceUsage, which will alter
maxMemoryUsage. After that, the value will never change, so it can
be read from multiple threads without any protection.