Bug fix 3.4/fix all numeric guids on restore (#11961)
* remove invalid GUID values when restoring collections from a dump

this fix allows restoring collections from v3.3.0 with their all-numeric
GUID values, and silently creates a new, valid GUID for them. v3.3.0
had a bug because it created all-numeric GUID values, which can be
confused with numeric collection ids in lookups. v3.3.1 changed the
GUID routine to produce something non-numeric already, but collections
created with v3.3.0 can have an ambiguous GUID. this patch adjusts
the restore routine to drop such GUIDs, so it only changes something
if the collections are dumped, dropped on the server and then restored
with the flawed GUIDs.

* updated CHANGELOG