Remove "lazyLoading"
according to Michael Hackstein this was put in to optimize how many
documents were loaded when using MMFiles; since we don't use MMFiles anymore
this is not needed anymore.

The code where it was attempted to be used was the ShortestPath algorithm
which crashed anyway (because of an unimplemented loadDocument function).