Avoid notifying condition variables while holding a lock. (#12206)
The lock is only required to serialize operations on the data the CV wait operations depends on. In some cases (e.g., when using atomics like in the ReadWriteLock) it is sufficient to lock and immediately unlock the mutex (i.e., an "empty critical secion") before calling notify.
Notifying a CV while holding the lock is a pessimization, because the notified thread will immediately try to acquire that lock and therefore be blocked again if the notifying thread has not yet released the lock.