Safeguards against using V8 when JavaScript is turned off
This PR adds safeguards against using V8 internally in environments that
have JavaScript turned off via the `--javascript.enabled false` option.

The PR also moves a bit of V8 context initialization code from the
Scheduler and the DatabaseFeature into the V8DealerFeature, so it is now
all in one place.

Finally it adds a new metrics that tracks the total time it took to
create V8 contexts.