take a copy before moving from the same object
the following code

    auth::User user = auth::User::fromDocument(s);
    result.try_emplace(user.username(), std::move(user));

moves from _user_ and accesses a member function in the same expression.

in practice, this should be safe, because the map's insert key
should be calculated first, and the value is only moved from if
the key does not exist in the map.

but I couldn't find any proof that this is indeed safe.
so changing it to an intentional and explicit copy to avoid
potential UB:

    auth::User user = auth::User::fromDocument(s);
    // intentional copy, as we are about to move-away from user
    std::string username = user.username();
    result.try_emplace(std::move(username), std::move(user));