make sure the x-arango-source header is set for all internal requests
There is currently no consequence from setting or not setting this
header except that there will be additional request logging in log level
DEBUG if the header is set. It is not used for anything else, but it may
help with debugging. Now it is set consistently for internal requests.