Fix LocalTaskQueue.
The _condition was never notified by stopped task. In addition, the termination criterion of the loop was incorrect, because it did not consider whether the queue still contained some tasks. So it could have happened that the loop was terminated prematurely, i.e., some tasks would have never been scheduled.