improve performance of shard synchronization protocol
by calling a special version of the inventory, which will only return
structural information about for the requested shard, and not for all
shards in the requested database. this can reduce the amount of data
transferred a lot in cases there are many shards present and they are
mostly in sync.