Adds `PRUNE` and array index variables to the `[* ...]` AQL operator
Note: this is experimental and may be deleted.

This is an RFC for an experimental change that adds `PRUNE` support to
the `[* ...]` array iterator operator in AQL. `PRUNE` will stop
iterating over the array as soon as its condition is *not* satisfied.
Note: `PRUNE` does not seem to be the ideal name for this functionality.
The programming-language analogy is `BREAK`. `PRUNE` is only used here
because we already use this word to achieve somewhat related behavior in
traversals.

The PR also adds an `INDEX` pseudo variable to the `[* ...]` AQL
operator. `INDEX` is a pseudo variable which contains the current loop
iteration index value (zero-based). The variable is made available
automatically inside `[* ...]` and is only valid inside. There already
exists a pseudo variable `CURRENT` to refer to the current array
element, and `INDEX` extends this.

A few usage examples:

```
LET a = [1, 2, 4, 16, 99, 257, 4, 10000]
RETURN a[* PRUNE INDEX > 0 && a[INDEX - 1] >= CURRENT RETURN CURRENT]

[
  1,
  2,
  4,
  16,
  99,
  257
]
```

```
LET a = [1, 2, 4, 8, 32, 64, 256]
RETURN a[* PRUNE INDEX > 0 && a[INDEX - 1] * 2 != CURRENT RETURN CURRENT]

[
  1,
  2,
  4,
  8
]
```

```
LET a = [1, 2, 4, 8, 32, 64, 256]
RETURN a[* PRUNE INDEX > 0 && a[INDEX - 1] * 2 != CURRENT RETURN INDEX]

[
  0,
  1,
  2,
  3
]
```

```
LET a = [1, 2, 4, 8, 32, 64, 256]
RETURN a[* RETURN {value: CURRENT, isDuplicate: INDEX == 0 || a[INDEX - 1] * 2 == CURRENT}]

[
  {
    "value" : 1,
    "isDuplicate" : true
  },
  {
    "value" : 2,
    "isDuplicate" : true
  },
  {
    "value" : 4,
    "isDuplicate" : true
  },
  {
    "value" : 8,
    "isDuplicate" : true
  },
  {
    "value" : 32,
    "isDuplicate" : false
  },
  {
    "value" : 64,
    "isDuplicate" : true
  },
  {
    "value" : 256,
    "isDuplicate" : false
  }
]
```

```
LET a = [1, 2, 4, 8, 16, 32, 64, 256, 512]
RETURN a[* FILTER INDEX % 2 == 1 RETURN CURRENT]

[
  2,
  8,
  32,
  256
]
```