remove multiple debug functions for producing backtraces.
Instead, make CrashHandler::logBacktrace() available to write a
backtrace of the current thread into the log.
Calling this function should work on Linux, because there we build with
libunwind support by default, and this can produce backtraces even for
our static builds using libmusl (in contrast to `backtrace()`, which is
a no-op there). In addition, using `CrashHandler::logBacktrace()` does
not require to compile with the `-DUSE_BACKTRACE=On` CMake option.

It is likely that the change made by this PR removes backtrace
capabilities for MacOS and Windows, but it looks like those were only
enabled previously when compiling with `-DUSE_BACKTRACE=On`, which is
unlikely.

To finalize this PR, we need to decide whether these potential
restrictions are ok for us, and if so, we can also remove the
`USE_BACKTRACE` CMake variable and remove the one or two remainders that
use the `ARNAGODB_ENABLE_BACKTRACE` define in the code.