Improve incremental sync replication
Improve incremental sync replication for single server and cluster to cope
with multiple secondary index unique constraint violations (before this
was limited to a failure in a single unique secondary index). this allows
replicating the leader state to the follower in basically any order, as
any *other* conflicting documents in unique secondary indexes will be
detected and removed on the follower.