Fix profiling of streaming AQL queries
Fix profiling of AQL queries with the `silent` and `stream` options sets in
combination. Using the `silent` option makes a query execute, but discard all
its results instantly. This led to some confusion in streaming queries, which
can return the first query results once they are available, but don't
necessarily execute the full query.
Now, `silent` correctly discards all results even in streaming queries, but
this has the effect that a streaming query will likely be executed completely
when the `silent` option is set. This is not the default however, and the
`silent` option is normally not set. There is no change for streaming queries
if the `silent` option is not set.