backport some changes from https://github.com/arangodb/arangodb/pull/13507 (#13520)
* backport some changes from https://github.com/arangodb/arangodb/pull/13507

This is a preparation PR for backporting some further replication tests
and bugfixes from devel to 3.7.
It isolates some of the changes done in
https://github.com/arangodb/arangodb/pull/13507 into a small PR to
facilitate reviewing. A further PR with the added replication tests and
a bugfix for 3.7 will follow up.

This PR extends the length of some internal byte buffers from 11 to 21,
so that they cannot overflow, regardless of what input is coming in. The
overflow was theoretically possible before, but only with manipulated
timestamp input.

The PR also fixes an issue in a RocksDB heap helper class that does a
potentially unsafe self-move assignment. This issue was raised when
running tests with a debug STL build (`_GLIBCXX_DEBUG`).

* Update CHANGELOG

Co-authored-by: Vadim <vadim@arangodb.com>