Small improvements (#14150)
* fix typo

* make an atomic change to internal state in takeOverLeadership,

and log an error message in case we are taking over leadership for a
shard that we haven't been a follower for before

* require at least 3.8 for sync-by-revisions protocol

* fail with an exception if we get an invalid pointer back

instead of potentially continuing with a nullptr

* added hidden option `--default-language-check`

so that the check between `--default-language` and the persisted
language value can be bypassed. the default value for this option is
`true`, meaning that the check will take place. it is not recommended to
turn it off except when debugging binary data from other installations.

* added CHANGELOG entry

Co-authored-by: Vadim <vadim@arangodb.com>