remember exception in ExecutionBlock
and return it on next invocation.
normally any exception returned by an ExecutionBlock is directly handled
by the block's caller.
however, if the same block is asked semi-concurrently (only protected by
the RestAqlHandler's lease), then there can still be other requests in
flight that wait to enter the block. if an exception escapes the block,
it may be left in an invalid state, so we have to prevent other threads
from entering it afterwards and continue with that invalid state.