try to improve unstable fillBlockCache test
For unknown reasons, the block cache usage values returned by the engine
stats in this test seem to be unstable, even in single server on Linux.
There is potentially some other, yet unknown action, that populates
parts of the RocksDB block cache.
Being unable to reproduce the issue locally, this is an attempt to cut a
part of the test in exchange for test stability.