fix known followers handling.
on sync replication failure, do not remove follower from known servers
in transaction.
if we do, we will not be able to send the commit/abort to the follower
later. However, we still need to send the commit/abort to the follower at
transaction end, because the follower may be responsbile  for _other_
shards as well.
it does not matter if we later commit the writes of the shard
from which we just removed the follower, because the follower
is now dropped and will try to get back in sync anyway, so
it will run the full shard synchronization process.