simplify blocker handling
this introduces RocksDBBlockerGuard so that acquiring and releasing
blockers is made simpler, and especially that releasing blockers is not
forgotten.
in addition, it makes placeBlocker() throw in case of an error. the
method previously returned a Result, but that was picked up nowhere.
throwing in case an (unlikely) error will force some sort of error
handling on the call site or an abort of the operation.