make statistics available in metrics API no matter what (#14702)
Previously, if `--server.statistics` was set to `false`, the metrics API
did not return any values provided by the StatisticsFeature, as the
statistics were turned off completely. This prevented the metrics API
from returning some useful process statistics (e.g. memory usage, number
of threads etc.).
This change now enables these process statistics as well as a few others
in metrics regardless of the setting of `--server.statistics`.

This was already the case in 3.8 and devel, but 3.7 still behaved
slightly differently

Co-authored-by: Vadim <vadim@arangodb.com>