ensure Query is created via shared_ptr (#14755)
* ensure Query is created via shared_ptr

This change makes sure that all instances of `Query` and `ClusterQuery`
are wrapped into `std::shared_ptr`s upon creation.
The constructors of both classes are now protected, meaning it is
impossible to create `Query` or `ClusterQuery` instance on the stack
from the outside.
The only remaining way to create a `Query` or `ClusterQuery` instance is
to call the classes' static `create` factory method, which will return a
`Query` or `ClusterQuery` object wrapped in a shared_ptr.

`Query` is now also derived from `std::enable_shared_from_this<Query>`,
meaning we can capture a query instance easily via calling
`shared_from_this()` later.

Apart from that, nothing is changed yet, so this PR does not yet solve
any problem. However, we want to capture shared_ptr-copies of `Query` in
callbacks and such, instead of capturing `SharedQueryState`. This should
have the advantage that the Query destructor will not be invoked until
all callbacks that have references to the query are finished.

* fix use-after-move

* simplification

* fix more use-after-moves

* fix use-after-move