Set default module root to cwd
This prevents accidental traversal to the file system root when attempting to resolve module paths while using arangosh. We already set the module root for Foxx services but left it unbounded by default so arangosh wouldn't know when to stop searching. This can have unexpected results when a node_modules folder is present at a higher level of hierarchy.