fix non-deterministic dropping of followers during tests (#15143)
* fix non-deterministic dropping of followers during tests

this change increases the supervision grace period from 10 seconds to a
higher value when running the `shell_client` tests. it seems that 10
seconds is not enough on overloaded servers, so that we randomly get
failedLeader jobs during the tests, which can lead to spurious dropping
of followers. this is a problem because some tests assume that during
the tests no followers are getting dropped.

* adjust possible status codes after change