do not copy generated log message string
when a log message is put together in an output stream, earlier versions
of C++ had no means of accessing the data from the ostream in a
read-only fashion without copying it. with c++20 it is now possible to
get a read-only view on the ostream's contents, without having to copy
it.
we can use this to avoid making a copy of the ostream content's for
every log message 😊