add optional "extraFields" attribute for non-unique persistent indexes
this will add the specified extra fields to the index, so they can be
used for projections (though not for lookups or sorting).

Example:

    db.<collection>.ensureIndex({ fields: ["value1"], type: "persistent", extraFields: ["value2"] });

This will index `value1` in the traditional sense, so the index can be
used for looking up by value1 or for sorting by value1. The index also
supports projections on value1.
Due to the `extraFields` being used here, the index can now also supply
the values for the `value2` attribute for projections.

This allows covering index scans in more cases and helps to avoid making
extra document lookups in the documents column family. This can have a
great effect on index scan performance if the number of index entries
looked at is large.