fix pitfall for comdat folding
we have two static methods with identical code, which returned the
addresses of the distinct method-local static objects. when enabling
comdat folding, the compiler decided to merge the two functions into
one, which gave both objects the same in-memory address.
this lead to follow-up errors because the object addresses were used to
tell two types of object (collections and views) apart.
this was not only error-prone, but also complicated. this PR replaces
all this with a simple enum class.