remove internal `_constAstNode` from `Variable` struct
the `_constAstNode` member in `Variable` was used before to track
the constant value of a variable in an AQL query. For example, in
`LET a = 1`, the `a` variable contained a `_constAstNode` pointer
to the AstNode containing `1`. This pointer was checked during
AST optimization and used for constant propagation, and remained
unused afterwards.
As we are tracking variable definitions during AST optimization in
a map anyway, this PR now removes the `_constAstNode` member from
`Variable` and simply relies on this map. This also allowed removal
of a few methods in `Ast` that became unused.