Improve restore parallelism for new data format (#16030)
In case the new data format (useEnvelope = false) was used with
arangodump, arangorestore can submit multiple batches with documents
concurrently. That way a restore can be multi-threaded even for a single
collection. Due to some invalid if condition, the logic was reversed and
parallelism was turned off when using the new format.

Co-authored-by: Vadim <vadim@arangodb.com>