decrease wait time in agency inception in case of HTTP 503
previously, when the inception received an HTTP 503 error, it could fail
with a VelocyPack error when trying to unserialize data in the response
body that was not there. This issue was fixed about a week ago. Now an
agent will handle 503 responses correctly, but it will go to a 10
seconds sleep. This is too long, because it delays the cluster start
unnecessarily. Reduce the sleep duration to 500ms when getting an HTTP
503.