fix broken tests
these tests were broken since the general test verbosity was reduced in
devel a few months ago. these tests check for specific log messages,
which were not emitted after the log verbosity was reduced. now, the
tests specifically enable the log topic in question, so that all
required log messages for the tests are emitted again.