adjust HTTP response code for error number 1521 from 500 to 400
Error 1521 (query collection lock failed) is nowadays only emitted in
traversals, when a collection is accessed during the traversal that has
not been specified in the WITH statement of the query. Thus returning
HTTP 500 is not a good idea, as it is clear a user error that triggered
the problem.