Stop total timer correctly when store=false
When store=false, Done state directly follows Running state without
cleaning up. Cleanup just happens when run is canceled, therefore the
total timer ran until the Pregel run was canceled and not - as it
should - until the computation is done.
Now the total timer stops directly when entering Done state and not storing.