fix an issue with --server.early-connections
when using `--server.early-connections`, the REST API is opened
early during startup and returns HTTP 200 for requests to `/_api/version`
and `/_admin/status`. this is as advertised, however, later in the
startup procedure the server will switch to maintenance mode, and then
it returned HTTP 503 for the same requests. this is inconsistent.
now, when using `--server.early-connections`, the server will keep
responding with HTTP 200 to the mentioned APIs during the entire
startup phase.