[3.9] Fix potential deadlock in rocksdb compaction. (#16598)
* Wraparound NextCompactionIndex in LevelCompactionBuilder

There can be situations where the picker is asked for a compaction twice
for the same version. In such a scenario the first call might not be able
to find any eligible files and as a result set NextCompactionIndex to the
total number of files. In this case the second call could return null even
though we could now compact one of the files we previously had to skip.
This could then result in a deadlock as explained in more detail here:
https://github.com/facebook/rocksdb/issues/10257#issuecomment-1180534048

* Update CHANGELOG

Co-authored-by: Vadim <vadim@arangodb.com>