deterministically sort shards in shards overview
Previously, the order of shards per collection was dependent on the
order of the shard name in an unsorted container, so it was
non-deterministic.
This led to problems in the "Shards" overview section of the UI, which
displayed the shards for a collection in the same order in which they
were stored in the server-side hash table. Now, we sort the shard IDs
in a determinstic way prior to returning them, which will make the
"Shards" overview in the UI display them in proper, unchanging order
without any shards "jumping" between reloads of the shard lists.