fix confusion about locking for _replicatedLogIds
TSan reported data races for accessing _replicatedLogIds. That container
was previously accessed under the _planProt lock, and sometimes under the
_currentProt lock.
This is now unified so that all accesses to _replicatedLogIds use the
_planProt lock.
Also the code for ClusterInfo::getResponsibleServer() contained some
special handling for replication2 shards, but it wasn't complete: while
the shards were detected correctly for replication2 cases, they were not
returned by the method, but the result for the replication1 case was
returned instead.