adjust permission checking of search-alias views
previously, "search-alias" views were visible to users that didn't have
read permissions on the underlying referenced collections. this was
inconsistent, because "arangosearch" views weren't shown to users that
didn't have read permissions on the underlying links.
now, the behavior for "search-alias" views is the same as for
"arangosearch" views, i.e. "search-alias" views are not shown and are
not accessible for users that don't have at least read permissions on
the underlying collections.