fix handling of parameters for `db._profileQuery(object)`
Previously, `db._profileQuery(...)` only accepted positional parameters.
This is in contrast to `db._query(...)` and `db._explain(...)`, which
both accept positional parameters and named parameters inside an object
input parameter.

With this PR, one can now do:

    db._profileQuery(queryString, bindVars, options);
as well as

    db._profileQuery({ query: queryString, bindVars: bindVars, options: options });

This PR also fixes the issue that profiling modifies the `options`
input parameter, so that when it is used for multiple calls, it is not
treated as const.