Make Geo faster:
* Make GeoJson parsing faster
* Make ShapeContainer intersects/contains faster
* Remove unnecessary functions
* Use FlatHashSet for deduplication uint64_t in query
* Use GeoJson parsing without validation for IResearch GeoFilter
* Fix centroid compute for S2LatLngRect in ShapeContainer
* Fix GEO_CONTAINS(S2Polygon, S2LatLngRect)
* Fix S2LatLng -> S2Point works only for valid S2LatLng
* Something else?
* Fix tests

Fix for empty polygons and multi polygons
Maybe we want to allow empty loops in general?
We can just skip such loops

If and return/throw error if S2LatLng not normalized

Allow to specify not normalized lat lon:

Before this commit it was really strange.
In some cases we produce error to user.
In other ignore and works incorrect.
In third makes normalize and works correct.
I think always produce error is better, but in such case we need way to update old user data.
It's impossible now, so I think better to always normalize

So we change behaviour:
In some queries user couldn't specify not normalized points.
It was error/invalid behavior, now they can and it will be valid