Allow to specify not normalized lat lon:
Before this commit it was really strange.
In some cases we produce error to user.
In other ignore and works incorrect.
In third makes normalize and works correct.
I think always produce error is better, but in such case we need way to update old user data.
It's impossible now, so I think better to always normalize

So we change behaviour:
In some queries user couldn't specify not normalized points.
It was error/invalid behavior, now they can and it will be valid