fix a potential race in sha file creation
a potential race can occur if an checkMissingShaFiles is called with a
requireAge value of 0. when then there is a missing .hash file for an
.sst file, first the current time was taken. then the file modification
time for the .sst file was retrieved, and the difference between the
current time and the file modification time was calculated.
a potential problem exists if the file is currently being modified and
its modification time is thus higher than the current time. then the
difference of the two times would be negative and not be >= requireAge.
in this case the hash value wouldn't be calculated and stored in a .hash
file.
this is at least theoretically possible. it is unclear if such situation
can occur in practice, because the function should not be called on .sst
files that are currently being modified. if the function is called on
.sst files are still being modified, there is also another issue that
the hashes are calculated and stored for partial files only.
this PR doesn't fix that, however, it doesn't make it any worse.